refactor: rename UserDetailsForm to UserFormFields and update imports in related components
